Bug that prevented GCode files being written when there was no RepRap plugged in fixed.

+++ reprap/host/src/org/reprap/Extruder.java
@@ -189,7 +189,7 @@ public interface Extruder
     /**
      * @return the cooling period in seconds
      */
-    public int getCoolingPeriod();
+    public double getCoolingPeriod();
  
     /**
      * @return the X offset in millimeters
